refactor: split SettingsView into focused components (#80)
SettingsView packed 676 lines: the screen itself, HealthKit integration
(with status state, refresh + handle helpers, and five status-display
computed properties), a DEBUG-only Developer section (with its own state
and confirmation dialogs), and a small UI kit (seven reusable Settings
component types) — four distinct concerns in one file.

Decomposed along the natural seams. Each extracted piece has its own state,
its own dependencies, and a clear single responsibility:

  SettingsComponents.swift     — Reusable settings UI primitives
                                 (SettingsRowStyle, SettingsIconTint,
                                 SettingsRow, SettingsRowIcon,
                                 SettingsSectionView, SettingsDivider,
                                 SegmentedToggle). These are now usable from
                                 anywhere, not just buried in one screen.
  SettingsHealthSection.swift  — Apple Health section. Owns its own
                                 healthKitStatus @State; queries it via
                                 .onAppear; encapsulates handleHealthRowTap
                                 and the five status-display helpers.
                                 Hidden on devices without HealthKit.
  SettingsDeveloperSection.swift — DEBUG-only Load/Clear test data tools.
                                 The whole file is #if DEBUG-gated. Owns
                                 its own showLoadConfirm / showClearConfirm
                                 @State and its two confirmationDialogs.

The smaller body sections (profile, units, appearance, data, about, reset)
stay inline in SettingsView — each is 13–50 lines of pure composition over
the shared components, with no per-section state. Extracting them would
just add init-parameter threading without making any concern simpler.

  SettingsView.swift: 676 → 270 lines (out of the oversized list)

Verified: make lint (0, --strict), make build, make test (272 logic +
13 UI incl. testSettingsScreenAccessibility, all green).
